+++ app/admin-new/admin.sass
@@ -113,6 +113,8 @@
 
     &-title
       @include body14M
+      color: $color-white50
+      
     &-value
       text-align: center
       color: $color-white
