+++ app/style/_variables.sass
@@ -24,7 +24,7 @@ $color-main: #131417
 $color-sub1: #1f2023
 $color-sub2: #3a3b3d
 $color-white5: #ffffff0d
-$color-white8: #ffffff1f
+$color-white8: #FFFFFF14
 $color-white12: #ffffff1f
 $color-white20: #ffffff33
 $color-white50: #ffffff80
