+++ stories/NewButton.component.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import "./_button.css";
+
+export function NewButton() {
+  return <button className="button32">hi</button>;
+}
