+++ stories/_button.css
@@ -1,0 +1,23 @@
+.button32 {
+  height: 32px;
+  padding: 10px 11px;
+  border-radius: 8px;
+  margin: auto;
+  font-size: 12px;
+  font-weight: 600;
+  line-height: 100%;
+}
+.button32:hover {
+  color: black;
+  background-color: rgba(255, 255, 255, 0.6980392157);
+}
+.button32:active {
+  color: black;
+  background-color: rgba(255, 255, 255, 0.5019607843);
+}
+.button32:disabled {
+  color: rgba(255, 255, 255, 0.2);
+  background-color: rgba(255, 255, 255, 0.2);
+}
+
+/*# sourceMappingURL=_button.css.map */
